+++ jarslink-api/src/main/java/com/alipay/jarslink/api/impl/RuntimeModule.java
@@ -26,6 +26,9 @@ public class RuntimeModule extends ToStringObject {
      */
     private String errorContext;
 
+    /**
+     * all version module,key:version
+     */
     private ConcurrentHashMap<String, Module> modules = new ConcurrentHashMap();
 
     public Module getModule(String version) {
